profile途中まで

+++ resources/views/posts/index.blade.php
@@ -9,12 +9,19 @@
   {{ Form::close() }}
 
   @foreach($posts as $post)
-    <a href="/user_profile/{{$post->user->id}}"><img src="{{ asset('images/' . $post->user->icon_image) }}" alt="ユーザーアイコン"></a>
+    @if(!$post->user_id != auth()->user()->id)
+      <a href="/user_profile/{{$post->user->id}}"><img src="{{ asset('images/' . $post->user->icon_image) }}" alt="ユーザーアイコン"></a>
+    @else
+      <a href="/profile"><img src="{{ asset('images/' . $post->user->icon_image) }}" alt="ユーザーアイコン"></a>
+    @endif
     {{$post->user->username}}
     {{$post->post}}
     {{$post->user->updated_at}}
     @if($post->user_id == auth()->user()->id)
-    <a href="/post/delete/{{$post->id}}"><img src="{{ asset('images/trash.png' ) }}" alt="削除ボタン"></a>
+      <a href="/post/update/{{$post->id}}"><img src="{{ asset('images/edit.png' ) }}" alt="編集ボタン"></a>
+      <a href="/post/delete/{{$post->id}}" onclick="return confirm('本当に削除しますか？')"><img src="{{ asset('images/trash.png' ) }}" alt="削除ボタン"></a>
+      @csrf
+      @method('DELETE')
     @endif
   @endforeach
 </x-login-layout>
